+++ Engine/src/engine/util/Sphere.cpp
@@ -161,7 +161,7 @@ void Sphere::clearArrays()
 ///////////////////////////////////////////////////////////////////////////////
 void Sphere::buildVerticesSmooth()
 {
-    const float PI = acos(-1);
+    const float PI = static_cast<float>(acos(-1));
 
     // clear memory of prev arrays
     clearArrays();
@@ -251,7 +251,7 @@ void Sphere::buildVerticesSmooth()
 ///////////////////////////////////////////////////////////////////////////////
 void Sphere::buildVerticesFlat()
 {
-    const float PI = acos(-1);
+    const float PI = static_cast<float>(acos(-1));
 
     // tmp vertex definition (x,y,z,s,t)
     struct Vertex
